[BEAM-7547] Avoid WindmillStateCache cache hits for stale work.
Add the work id to the cache and use the sequencing of work ids to
prevent reading a cached value for stale work items.  This prevents
viewing cached state that is inconsistent with the work.  Reads of
persistent state to the Dataflow backend will fail with invalid work
token for stale work.